+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - jxlazy: allow querying extra channel blend info.
 - jxlazy: new `unbounded` indicator to identify boxes that extend to the end of the file.
+- `add` and `subtract` command line modes.
 
 ### Fixed